Update Similarität_Preprocessing.R
Calculates Distance between a cell of a DF and the cell below comtaining strings. Adds new column with the distance for each cell. It adds a roa called SimSum that enables to see the context above and below of each row with a certain threshold. This faciliates preprocessing of corpus data. Filter SimSum column in a calc program by > 0 in order to see similar cells with their context.